+++ core/test/roms/tests.js
@@ -36,7 +36,7 @@ describe('Validation ROMs', () => {
   validate(ppu_vbl_nmi);
   validate(apu_reset); // 4 failing tests (disabled)
   validate(apu_test); // 3 failing tests (disabled)
-  validate(mmc3_test); // 2 failing tests (disabled)
+  validate(mmc3_test); // 1 failing tests (disabled)
   validate(bntest);
   validate(holydiverbatman);
 });